# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Postman-like API testing platform. Authenticate with Bearer JWT for /api/* routes, or X-API-Key for /api/v1/* routes.",
        "title": "Postmanxodja API",
        "termsOfService": "https://postbaby.uz",
        "contact": {
            "name": "Support",
            "url": "https://postbaby.uz"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/collections": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List collections using an API key. Useful for CI/CD and third-party integrations.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public-api"
                ],
                "summary": "List collections (API key)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Collection"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Accepts either {\"raw_json\":\"...\"}, {\"name\":\"...\",\"description\":\"...\"}, or a raw Postman collection JSON body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public-api"
                ],
                "summary": "Create collection (API key)",
                "parameters": [
                    {
                        "description": "Collection payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Collection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/collections/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public-api"
                ],
                "summary": "Get collection (API key)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Collection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "collection": {
                                    "$ref": "#/definitions/models.PostmanCollection"
                                },
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public-api"
                ],
                "summary": "Update collection (API key)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Collection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New collection JSON",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "raw_json": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Collection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public-api"
                ],
                "summary": "Delete collection (API key)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Collection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Collection": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "environment_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "raw_json": {
                    "type": "string"
                },
                "team_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.PostmanAuth": {
            "type": "object",
            "properties": {
                "apikey": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostmanAuthParameter"
                    }
                },
                "basic": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostmanAuthParameter"
                    }
                },
                "bearer": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostmanAuthParameter"
                    }
                },
                "oauth2": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostmanAuthParameter"
                    }
                },
                "type": {
                    "description": "bearer, basic, apikey, oauth2, etc.",
                    "type": "string"
                }
            }
        },
        "models.PostmanAuthParameter": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "models.PostmanBodyOptions": {
            "type": "object",
            "properties": {
                "raw": {
                    "$ref": "#/definitions/models.PostmanRawOptions"
                }
            }
        },
        "models.PostmanCollection": {
            "type": "object",
            "properties": {
                "info": {
                    "$ref": "#/definitions/models.PostmanInfo"
                },
                "item": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostmanItem"
                    }
                },
                "variable": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostmanVariable"
                    }
                }
            }
        },
        "models.PostmanFormData": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "disabled": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "type": {
                    "description": "text, file",
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.PostmanInfo": {
            "type": "object",
            "properties": {
                "_collection_link": {
                    "type": "string"
                },
                "_exporter_id": {
                    "type": "string"
                },
                "_postman_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "schema": {
                    "type": "string"
                }
            }
        },
        "models.PostmanItem": {
            "type": "object",
            "properties": {
                "item": {
                    "description": "For folders",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostmanItem"
                    }
                },
                "name": {
                    "type": "string"
                },
                "request": {
                    "$ref": "#/definitions/models.PostmanRequest"
                },
                "response": {
                    "description": "Saved example responses",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostmanResponse"
                    }
                }
            }
        },
        "models.PostmanKeyValue": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "disabled": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "models.PostmanRawOptions": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string"
                }
            }
        },
        "models.PostmanRequest": {
            "type": "object",
            "properties": {
                "auth": {
                    "$ref": "#/definitions/models.PostmanAuth"
                },
                "body": {
                    "$ref": "#/definitions/models.PostmanRequestBody"
                },
                "header": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostmanKeyValue"
                    }
                },
                "method": {
                    "type": "string"
                },
                "url": {
                    "description": "Can be string or PostmanURL"
                }
            }
        },
        "models.PostmanRequestBody": {
            "type": "object",
            "properties": {
                "formdata": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostmanFormData"
                    }
                },
                "mode": {
                    "description": "raw, formdata, urlencoded",
                    "type": "string"
                },
                "options": {
                    "$ref": "#/definitions/models.PostmanBodyOptions"
                },
                "raw": {
                    "type": "string"
                },
                "urlencoded": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostmanUrlEncoded"
                    }
                }
            }
        },
        "models.PostmanResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "header": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PostmanKeyValue"
                    }
                },
                "name": {
                    "type": "string"
                },
                "originalRequest": {
                    "$ref": "#/definitions/models.PostmanRequest"
                },
                "responseTime": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.PostmanUrlEncoded": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "disabled": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.PostmanVariable": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Team API key (prefix pmx_).",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}